Separate slices and thunks
